requirements.txt
djangorestframework-3.10.2
